Move GameDatasetInfo to own file, renamed to DatasetInfo.
